refactor: Change PriceType
price, totalPrice 타입 변경
 - int -> long

+++ src/main/java/com/example/settlementnew/api/BankApi.java
@@ -2,5 +2,5 @@ package com.example.settlementnew.api;
 
 public interface BankApi {
 
-    History transfer(String fromUsername, String toUsername, int price) throws Exception;
+    History transfer(String fromUsername, String toUsername, long price) throws Exception;
 }
