+++ src/main/java/com/example/settlementnew/api/DefaultBankApi.java
@@ -11,7 +11,7 @@ public class DefaultBankApi implements BankApi {
     private final double CRITICAL_ERROR_RATE = 0.0002;
 
     @Override
-    public History transfer(String fromUsername, String toUsername, int price) throws Exception {
+    public History transfer(String fromUsername, String toUsername, long price) throws Exception {
         double random = Math.random();
         if (random < CRITICAL_ERROR_RATE) {
             fatalError(toUsername, price);
@@ -21,12 +21,12 @@ public class DefaultBankApi implements BankApi {
         return new History(true, fromUsername, toUsername, price, LocalDateTime.now());
     }
 
-    private static History reasonableError(String from, String to, int price) {
+    private static History reasonableError(String from, String to, long price) {
         log.warn("송금 중 에러 발생 to: {}, amount: {}", to, price);
         return new History(false, from, to, price);
     }
 
-    private static void fatalError(String to, int price) throws Exception {
+    private static void fatalError(String to, long price) throws Exception {
         log.error("송금 중 치명적 에러 발생 to: {}, amount: {}", to, price);
         throw new Exception("치명적 에러 발생");
     }
