+++ src/main/java/com/example/settlementnew/api/History.java
@@ -15,11 +15,11 @@ public class History {
 
     private String to;
 
-    private int amount;
+    private long amount;
 
     private LocalDateTime createdAt;
 
-    public History(boolean status, String from, String to, int amount) {
+    public History(boolean status, String from, String to, long amount) {
         this.status = status;
         this.from = from;
         this.to = to;
@@ -27,7 +27,7 @@ public class History {
         this.createdAt = LocalDateTime.now();
     }
 
-    public History(boolean status, String from, String to, int amount, LocalDateTime createdAt) {
+    public History(boolean status, String from, String to, long amount, LocalDateTime createdAt) {
         this.status = status;
         this.from = from;
         this.to = to;
